--- README.txt
@@ -1,23 +1,0 @@
-# Thimble: Your Secure Messaging Solution 
-
-## Overview 
-
-Thimble is a secure, military-grade messaging system proof-of-concept that 
-prioritizes simplicity, security, and usability. It's not a silver bullet for 
-all communication needs, but a focused tool designed to enhance your specific 
-workflows.
-
-## Key Benefits
-
-* **Security:** End-to-end encryption ensures your messages remain confidential.
-* **Reliability:** Message history is stored in a Git repository for easy auditing and backup.
-* **Efficiency:** A streamlined interface helps you communicate quickly and effectively.
-
-## Why Thimble?
-
-Thimble fills a unique niche. It's perfect for situations where secure, trackable messaging is essential, 
-but where the full features of a traditional email client aren't necessary.
-
-## Getting Started
-
-[Instructions on how to install and use Thimble]
