+++ src/js/generate_report.js
